refactor(paths): move insight to its own folder
