+++ posthog/hogql_queries/insights/paths/test/test_paths_query_runner.py
@@ -15,7 +15,7 @@ from dateutil.relativedelta import relativedelta
 
 from posthog.schema import CachedPathsQueryResponse
 
-from posthog.hogql_queries.insights.paths_query_runner import PathsQueryRunner
+from posthog.hogql_queries.insights.paths.paths_query_runner import PathsQueryRunner
 from posthog.models import Team
